+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## December 11, 2017
+
+- Modify process-TMB-spreadsheets R scripts
+
 ## December 5, 2017
 
 - Initial commit of R codes